requirements.txt
arrow==0.12.1
atomicwrites==1.2.1
attrs==18.2.0
binaryornot==0.4.4
bumpversion==0.5.3
certifi==2018.8.24
chardet==3.0.4
click==6.7
cookiecutter==1.6.0
flake8==3.5.0
future==0.16.0
idna==2.7
Jinja2==2.10
jinja2-time==0.2.0
MarkupSafe==1.0
mccabe==0.6.1
more-itertools==4.3.0
pathlib2==2.3.2
pkg-resources==0.0.0
pkginfo==1.4.2
pluggy==0.7.1
poyo==0.4.1
py==1.6.0
pycodestyle==2.3.1
pyflakes==1.6.0
pytest==3.8.0
python-dateutil==2.7.3
PyYAML==3.13
requests==2.19.1
requests-toolbelt==0.8.0
six==1.11.0
toml==0.9.6
tox==3.3.0
tqdm==4.26.0
twine==1.11.0
urllib3==1.23
virtualenv==16.0.0
whichcraft==0.4.1
